# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Testing Swagger APIs.",
        "title": "Exam Data Collector API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/available_exams/create_available_exam": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create a new available exam in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Available_Exams"
                ],
                "summary": "Create a new available exam",
                "parameters": [
                    {
                        "description": "Available_Exams",
                        "name": "available_exam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateAvailableExamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AvailableExams"
                        }
                    }
                }
            }
        },
        "/api/available_exams/get_all_available_exams": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get a list of available exam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Available_Exams"
                ],
                "summary": "Get all available exams",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AvailableExams"
                            }
                        }
                    }
                }
            }
        },
        "/api/available_exams/{id}/delete_available_exam": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Delete a available exam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Available_Exams"
                ],
                "summary": "Delete a available exam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Available Exam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Available exam data deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/available_exams/{id}/get_available_exam": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get available exam detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Available_Exams"
                ],
                "summary": "Get available exam",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Available Exam ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AvailableExams"
                        }
                    }
                }
            }
        },
        "/api/customers/create_customer": {
            "post": {
                "description": "Create a new user in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/api/customers/get_all_customers": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get a list of users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/{id}/delete_customer": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Delete a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/customers/{id}/get_customer": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/api/customers/{id}/update_customer": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "User update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AvailableExams": {
            "type": "object",
            "required": [
                "dates",
                "exam_type",
                "months"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "exam_type": {
                    "$ref": "#/definitions/models.ExamTypeDetails"
                },
                "months": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateAvailableExamRequest": {
            "type": "object",
            "required": [
                "dates",
                "exam_type",
                "months"
            ],
            "properties": {
                "dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "exam_type": {
                    "type": "string"
                },
                "months": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.CreateUserRequest": {
            "type": "object",
            "required": [
                "day",
                "dob",
                "email",
                "exam_id",
                "month",
                "name",
                "prometric_id",
                "prometric_password"
            ],
            "properties": {
                "day": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "exam_id": {
                    "type": "string"
                },
                "month": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prometric_id": {
                    "type": "string"
                },
                "prometric_password": {
                    "type": "string"
                }
            }
        },
...
        "models.ExamTypeDetails": {
            "type": "object",
            "required": [
                "exam_type"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "exam_type": {
                    "type": "string"
                },
                "section": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.UpdateUserRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "required": [
                "day",
                "dob",
                "email",
                "exam_id",
                "month",
                "name",
                "prometric_id",
                "prometric_password",
                "status"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "day": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "exam_id": {
                    "type": "string"
                },
                "month": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prometric_id": {
                    "type": "string"
                },
                "prometric_password": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}